Icon: Remove %f in launch command
The '%f' appended to the launch command in the `.app` file is not
interpreted correctly by the dashboard. It is utilised by pcmanfm to
launch Sonic Pi with a file but with the transition away from classic
mode towards the dashboard, fixing it in the dashboard is preferable.
The solution here is to simply remove the string.